move pre-defined tags to pytato.tags
